Neovim: Remove Mason LSP manager
This is definetly a downside of NixOs.
Because of the way binaries are handled on nix, it is not possible to
use Mason to manage and isntall lsps. Thankfully doing it by hand is not
that complicated, at least not for the few languages I have done so far.

Nix: nixd - installed globally
Lua: lua-language-server lua_ls - installed globally
Rust: rust_analyzer - installed only in dev environments via flake-shell