Fix full boot partition + Update Flake
Right now I am deleting all old boot partitions daily. Not sure if that
is smart in the long run, but it should at least solve the isseu of
running out of space on /boot.

The issue is that nixos-rebuild boot tries to install the new kernel
before deleting old kernels. I only have space for a few kernels,
because of Windows dual boot, so if there are to many generations with
different kernel images I run out of space and can no longer run
nixos-rebuild boot (implicitly called by switch).
The solution in that case is to manually delete some old kernel image,
preferely the oldest and rerun nixos-rebuild boot.
However that should no longer be necessary.